+++ src/framework/piral-core/src/modules/api.ts
@@ -66,7 +66,7 @@ export function mergeApis(api: PiletApi, extenders: Array<PiletApiExtender<Parti
 }
 
 export function createExtenders(context: GlobalStateContext, apis: Array<PiralPlugin>) {
-  const creators: Array<PiralPlugin> = [...apis.filter(isfunc), createCoreApi];
+  const creators: Array<PiralPlugin> = [createCoreApi, ...apis.filter(isfunc)];
   return creators.map((c) => {
     const ctx = c(context);
 
@@ -80,7 +80,7 @@ export function createExtenders(context: GlobalStateContext, apis: Array<PiralPl
   });
 }
 
-export function defaultApiCreator(context: GlobalStateContext, apis: Array<PiralPlugin>): PiletApiCreator {
+export function defaultApiFactory(context: GlobalStateContext, apis: Array<PiralPlugin>): PiletApiCreator {
   const extenders = createExtenders(context, apis);
   return (target) => {
     const api = initializeApi(target, context);